Broken states

+++ src/main/java/br/com/ifpr/edu/sdpe_backend/service/ProjetoService.java
@@ -1,7 +1,7 @@
 package br.com.ifpr.edu.sdpe_backend.service;
 
 import br.com.ifpr.edu.sdpe_backend.domain.*;
-//import br.com.ifpr.edu.sdpe_backend.domain.DTO.EstatisticaDTO;
+import br.com.ifpr.edu.sdpe_backend.repository.ContatoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.InstituicaoEnsinoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.ProjetoRepository;
 import br.com.ifpr.edu.sdpe_backend.repository.VisualizacaoRepository;
@@ -37,6 +37,8 @@ public class ProjetoService {
 
     private final VisualizacaoRepository visualizacaoRepository;
 
+    private final ContatoRepository contatoRepository;
+
     private final Path rootLocation = Paths.get("uploads");
 
     public Projeto salvar(Projeto projeto) {
@@ -45,7 +47,7 @@ public class ProjetoService {
 
     public Projeto salvar(Projeto projeto, MultipartFile arquivo, String emailCoordenador) throws IOException {
 
-        projeto.setStatus(true);
+        projeto.setAtivo(false);
 
         if (projeto.getDataInicio() != null && projeto.getDataFim() != null) {
             if (projeto.getDataFim().before(projeto.getDataInicio())) {
@@ -99,7 +101,6 @@ public class ProjetoService {
 
 
                 if (busca.isPresent()) {
-                    // Se já existe, usamos a do banco (ignorando a descrição nova para não sobrescrever dados antigos sem querer)
                     instituicaoFinal = busca.get();
                 } else {
                     instituicaoFinal = instituicaoEnsinoRepository.save(input);
@@ -120,7 +121,7 @@ public class ProjetoService {
 
     public Page<Projeto> buscarTodos(int numPag, int tamPag) {
         Pageable pageable = PageRequest.of(numPag, tamPag);
-        return this.projetoRepository.findAll(pageable);
+        return this.projetoRepository.findByAtivo(true, pageable);
     }
 
     public Projeto atualizar(Projeto projeto, Long id, MultipartFile imagem) throws IOException {
@@ -135,9 +136,11 @@ public class ProjetoService {
         existente.setCargaHoraria(projeto.getCargaHoraria());
         existente.setFormato(projeto.getFormato());
 
-        if (projeto.getRedesSociais() != null) {
-            existente.setRedesSociais(projeto.getRedesSociais());
-        }
+        existente.setAtivo(false);
+        existente.setMotivoRejeicao(null);
+
+        if (projeto.getRedesSociais() != null) existente.setRedesSociais(projeto.getRedesSociais());
+
 
         if (projeto.getInstituicaoEnsino() != null) {
             InstituicaoEnsino inst = instituicaoEnsinoRepository.findByNome(projeto.getInstituicaoEnsino().getNome())
@@ -164,26 +167,34 @@ public class ProjetoService {
         return existente;
     }
 
+    @Transactional
     public void excluir(Long id) {
-        Projeto projeto = this.projetoRepository.findById(id).orElseThrow(
-                () -> new EntityNotFoundException("Projeto a ser atualizado não encontrado"));
+        Projeto projeto = this.buscarPorId(id);
 
-        if (projeto.getImagemPath() != null && !projeto.getImagemPath().trim().isEmpty())
-            deletarArquivoFisico(projeto.getImagemPath());
+        visualizacaoRepository.deleteByProjetoId(id);
 
-        if (projeto.getDocumentoPath() != null && !projeto.getDocumentoPath().trim().isEmpty())
-            deletarArquivoFisico(projeto.getDocumentoPath());
+        List<Contato> contatos = contatoRepository.findByProjetoId(id);
+        if (!contatos.isEmpty()) contatoRepository.deleteAll(contatos);
+
+        deletarArquivoFisico(projeto.getImagemPath());
+        deletarArquivoFisico(projeto.getDocumentoPath());
 
         this.projetoRepository.deleteById(id);
     }
 
-    private void deletarArquivoFisico(String nomeArquivo) {
-        try {
-            Path arquivo = rootLocation.resolve(nomeArquivo);
+    private void deletarArquivoFisico(String caminhoOuUrl) {
+        if (caminhoOuUrl != null && !caminhoOuUrl.isBlank()) {
+            try {
+                String nomeArquivo = caminhoOuUrl;
+                if (caminhoOuUrl.contains("/")) {
+                    nomeArquivo = caminhoOuUrl.substring(caminhoOuUrl.lastIndexOf("/") + 1);
+                }
 
-            Files.deleteIfExists(arquivo);
-        } catch (IOException e) {
-            System.err.println("Erro ao deletar arquivo físico (" + nomeArquivo + "): " + e.getMessage());
+                Path arquivo = rootLocation.resolve(nomeArquivo);
+                Files.deleteIfExists(arquivo);
+            } catch (Exception e) {
+                System.err.println("Erro ao deletar arquivo físico (" + caminhoOuUrl + "): " + e.getMessage());
+            }
         }
     }
 
@@ -254,9 +265,16 @@ public class ProjetoService {
         }
     }
 
+    public List<Contato> listarFeedbacks(Long idProjeto) {
+        buscarPorId(idProjeto);
+        return contatoService.buscarPorProjeto(idProjeto);
+    }
+
     public Contato adicionarFeedback(Long idProjeto, Contato contato) {
         Projeto projeto = buscarPorId(idProjeto);
+
         contato.setProjeto(projeto);
+
         return contatoService.salvar(contato);
     }
